refector: regions organize package
