+++ src/main/java/com/rgv04/hr/api/controller/RegionController.java
@@ -10,13 +10,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.rgv04.hr.domain.region.Region;
-import com.rgv04.hr.domain.region.RegionFilter;
-import com.rgv04.hr.domain.region.RegionService;
-import com.rgv04.hr.domain.region.assembler.RegionAssembler;
-import com.rgv04.hr.domain.region.assembler.RegionWithCountryAssembler;
-import com.rgv04.hr.domain.region.assembler.model.RegionModel;
-import com.rgv04.hr.domain.region.assembler.model.RegionWithCountryModel;
+import com.rgv04.hr.domain.assembler.RegionAssembler;
+import com.rgv04.hr.domain.assembler.RegionWithCountryAssembler;
+import com.rgv04.hr.domain.dto.RegionFilter;
+import com.rgv04.hr.domain.dto.RegionModel;
+import com.rgv04.hr.domain.dto.RegionWithCountryModel;
+import com.rgv04.hr.domain.model.Region;
+import com.rgv04.hr.domain.service.RegionService;
 
 @RestController
 @RequestMapping("api/regions")
