+++ src/main/java/com/rgv04/hr/domain/dto/EmployeeModel.java
@@ -11,6 +11,7 @@ import com.rgv04.hr.domain.job.controller.model.JobModel;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -20,9 +21,11 @@ import lombok.Setter;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
+@EqualsAndHashCode(callSuper = false)
 @Relation(collectionRelation = "employees")
 public class EmployeeModel extends RepresentationModel<EmployeeModel> {
 
+	@EqualsAndHashCode.Include
 	private Long id;
 
 	private String firstName;
