+++ src/main/java/com/rgv04/hr/domain/dto/RegionFilter.java
@@ -1,4 +1,4 @@
-package com.rgv04.hr.domain.region;
+package com.rgv04.hr.domain.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
