+++ src/main/java/com/rgv04/hr/domain/dto/RegionModel.java
@@ -1,10 +1,11 @@
-package com.rgv04.hr.domain.region.assembler.model;
+package com.rgv04.hr.domain.dto;
 
 import org.springframework.hateoas.RepresentationModel;
 import org.springframework.hateoas.server.core.Relation;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -14,9 +15,11 @@ import lombok.Setter;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
+@EqualsAndHashCode(callSuper = false)
 @Relation(collectionRelation = "regions")
 public class RegionModel extends RepresentationModel<RegionModel> {
 
+	@EqualsAndHashCode.Include
 	private Long id;
 
 	private String name;
