+++ src/main/java/com/rgv04/hr/domain/dto/RegionWithCountryModel.java
@@ -1,20 +1,20 @@
-package com.rgv04.hr.domain.region.assembler.model;
+package com.rgv04.hr.domain.dto;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.hateoas.server.core.Relation;
 
-import com.rgv04.hr.domain.dto.CountryModel;
-
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
+@EqualsAndHashCode(callSuper = false)
 @Relation(collectionRelation = "regions-with-countries")
 public class RegionWithCountryModel extends RegionModel {
 
-    private List<CountryModel> countries = new ArrayList<CountryModel>();
+    private List<CountryModel> countries = new ArrayList<>();
 
 }
