+++ src/main/java/com/rgv04/hr/domain/service/CountryService.java
@@ -10,7 +10,7 @@ import com.rgv04.hr.domain.dto.CountryFilter;
 import com.rgv04.hr.domain.dto.CountryModel;
 import com.rgv04.hr.domain.exception.CountryNotFoundException;
 import com.rgv04.hr.domain.model.Country;
-import com.rgv04.hr.domain.region.Region;
+import com.rgv04.hr.domain.model.Region;
 import com.rgv04.hr.domain.repository.CountryRepository;
 
 import lombok.RequiredArgsConstructor;
