+++ src/main/java/com/rgv04/hr/domain/service/RegionService.java
@@ -1,4 +1,4 @@
-package com.rgv04.hr.domain.region;
+package com.rgv04.hr.domain.service;
 
 import java.util.List;
 
@@ -6,6 +6,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.rgv04.hr.domain.dto.RegionFilter;
+import com.rgv04.hr.domain.exception.RegionNotFoundException;
+import com.rgv04.hr.domain.model.Region;
+import com.rgv04.hr.domain.repository.RegionRepository;
 import com.rgv04.hr.exception.BusinessException;
 
 @Service
