+++ src/main/java/com/rgv04/hr/infrastructure/RepresentationModelAssemblerSupport.java
@@ -1,4 +1,4 @@
-package com.rgv04.hr.domain.region.assembler;
+package com.rgv04.hr.infrastructure;
 
 public interface RepresentationModelAssemblerSupport<T1, T2> {
